Update eTrack
Tried to link it with the workflow. I put the paths to the assembly fastas in the way you have done above. I also changed it from clustalw to mafft, because clustalw takes ages to run. That should do what we want for phylogenetics. The only problem is that our environment doesn't have mafft. If it is unnecessary then you can delete mafft but i found it much more efficient and less time consuming.