Update eTrack
Tried to link to the workflow of the script and added the path to the assembly fastas as stated above and also changed the alignment to mafft which is much faster than clustalw. The last code which i run on Saturday morning is still running even today using clustalw. I remembered you suggested mafft and it is much faster so far. Though mafft is not in our environment. If it is not necessary feel free to remove it. I just commented out the clustalw code.