Update eTrack
I have added the code that should work. I hushed out the code for whole genome alignment so you can decide whether to keep it or remove it. You will check and see whether the paths i have included make sense and  the code in general. But it should run. It produces the newick file as a .tree file. I will send you the folder of the output from my run on slack.